Create SpectacleClient and move command generation calls there
