+++ workspaces/ui-v2/src/clients/index.ts
@@ -1,0 +1,1 @@
+export * from './spectacle';
