+++ workspaces/ui-v2/src/pages/docs/components/SimulatedBodyPreview.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 
 import { CQRSCommand } from '@useoptic/optic-domain';
 
@@ -6,6 +6,7 @@ import { ShapeFetcher, SimulatedCommandStore } from '<src>/components';
 import { useSpectacleContext } from '<src>/contexts/spectacle-provider';
 import { useAppSelector } from '<src>/store';
 import { IShapeRenderer } from '<src>/types';
+import { SpectacleClient } from '<src>/clients';
 
 type SimulatedBodyProps = {
   rootShapeId: string;
@@ -19,6 +20,9 @@ export const SimulatedBody = ({
   children,
 }: SimulatedBodyProps) => {
   const spectacle = useSpectacleContext();
+  const spectacleClient = useMemo(() => new SpectacleClient(spectacle), [
+    spectacle,
+  ]);
   const [previewCommands, setPreviewCommands] = useState<CQRSCommand[]>([]);
   // Commands that generate commands - removed fields
   const removedFields = useAppSelector(
@@ -30,16 +34,9 @@ export const SimulatedBody = ({
     (async () => {
       // TODO FLEB add in edited fields
       const removeFieldCommandsPromise: Promise<CQRSCommand[]> = Promise.all(
-        removedFields.map((fieldId) => {
-          // TODO FLEB generate removed field id via spectacle
-          return Promise.resolve([
-            {
-              RemoveField: {
-                fieldId,
-              },
-            },
-          ]);
-        })
+        removedFields.map((fieldId) =>
+          spectacleClient.fetchFieldRemoveCommands(fieldId)
+        )
       ).then((fieldCommands) => fieldCommands.flat());
 
       const [removeFieldCommands] = await Promise.all([
@@ -54,7 +51,7 @@ export const SimulatedBody = ({
     return () => {
       isStale = true;
     };
-  }, [removedFields]);
+  }, [removedFields, spectacleClient]);
 
   return (
     <SimulatedCommandStore
