+++ workspaces/ui-v2/src/store/documentationEdit/thunks.ts
@@ -8,53 +8,18 @@ import {
   CQRSCommand,
   PrunePathComponents,
 } from '@useoptic/optic-domain';
+import { SpectacleClient } from '<src>/clients';
 import { RootState, AppDispatch } from '../root';
 import { getValidContributions } from '../selectors';
-
-type EndpointProjection = {
-  endpoint: {
-    commands: {
-      remove: CQRSCommand[];
-    };
-  } | null;
-};
 
 const fetchRemoveEndpointCommands = async (
   spectacle: IForkableSpectacle,
   pathId: string,
   method: string
 ): Promise<CQRSCommand[]> => {
+  const spectacleClient = new SpectacleClient(spectacle);
   try {
-    const results = await spectacle.query<
-      EndpointProjection,
-      {
-        pathId: string;
-        method: string;
-      }
-    >({
-      query: `
-      query X($pathId: ID!, $method: String!) {
-        endpoint(pathId: $pathId, method: $method) {
-          commands {
-            remove
-          }
-        }
-      }`,
-      variables: {
-        pathId,
-        method,
-      },
-    });
-    if (results.errors) {
-      console.error(results.errors);
-      throw new Error(JSON.stringify(results.errors));
-    }
-    if (!results.data || !results.data.endpoint) {
-      const message = `Could not generate removal commands for endpoint path: ${pathId} and method: ${method}`;
-      console.error(message);
-      throw new Error(message);
-    }
-    return results.data.endpoint.commands.remove;
+    return spectacleClient.fetchRemoveEndpointCommands(pathId, method);
   } catch (e) {
     console.error(e);
     Sentry.captureException(e);
@@ -62,47 +27,13 @@ const fetchRemoveEndpointCommands = async (
   }
 };
 
-type FieldCommands = {
-  field: {
-    commands: {
-      remove: CQRSCommand[];
-    };
-  } | null;
-};
-
 const fetchFieldRemoveCommands = async (
   spectacle: IForkableSpectacle,
   fieldId: string
 ): Promise<CQRSCommand[]> => {
+  const spectacleClient = new SpectacleClient(spectacle);
   try {
-    const results = await spectacle.query<
-      FieldCommands,
-      {
-        fieldId: string;
-      }
-    >({
-      query: `
-      query X($fieldId: ID!) {
-        field(fieldId: $fieldId) {
-          commands {
-            remove
-          }
-        }
-      }`,
-      variables: {
-        fieldId,
-      },
-    });
-    if (results.errors) {
-      console.error(results.errors);
-      throw new Error(JSON.stringify(results.errors));
-    }
-    if (!results.data || !results.data.field) {
-      const message = `Could not generate removal commands for field: ${fieldId}`;
-      console.error(message);
-      throw new Error(message);
-    }
-    return results.data.field.commands.remove;
+    return spectacleClient.fetchFieldRemoveCommands(fieldId);
   } catch (e) {
     console.error(e);
     Sentry.captureException(e);
